backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=24.1
psycopg2-binary>=2.9
python-decouple>=3.8
dj-database-url>=2.1
gunicorn>=22.0
whitenoise>=6.6
drf-spectacular>=0.27
